refactor: app module to presentation rename
