+++ presentation/src/main/java/wolf/shin/presentation/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package wolf.shin.androidearlyadopter.ui.theme
+package wolf.shin.presentation.ui.theme
 
 import androidx.compose.ui.graphics.Color
