+++ presentation/src/main/java/wolf/shin/presentation/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package wolf.shin.androidearlyadopter.ui.theme
+package wolf.shin.presentation.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
